+++ apps/prairielearn/assets/scripts/question.ts
@@ -21,6 +21,25 @@ onDocumentReady(() => {
   disableOnSubmit();
 
   $('.js-submission-body.render-pending').on('show.bs.collapse', loadPendingSubmissionPanel);
+
+  const copyQuestionForm = document.querySelector<HTMLFormElement>('#copyQuestionModal form');
+  if (copyQuestionForm) {
+    const courseSelect = copyQuestionForm.querySelector<HTMLSelectElement>(
+      '#copyQuestionModal select[name="to_course_id"]',
+    );
+    courseSelect?.addEventListener('change', () => {
+      const option = courseSelect.selectedOptions[0];
+
+      if (option) {
+        copyQuestionForm.action = option?.dataset.copyUrl ?? '';
+        copyQuestionForm
+          .querySelectorAll<HTMLInputElement>('input[name="__csrf_token"]')
+          .forEach((input) => {
+            input.value = option?.dataset.csrfToken ?? '';
+          });
+      }
+    });
+  }
 });
 
 function externalGradingLiveUpdate() {
@@ -95,6 +114,7 @@ function fetchResults(socket: Socket, submissionId: string) {
     variantId,
     questionId,
     instanceQuestionId,
+    userId,
     variantToken,
     urlPrefix,
     questionContext,
@@ -112,6 +132,7 @@ function fetchResults(socket: Socket, submissionId: string) {
       question_id: questionId,
       instance_question_id: instanceQuestionId === '' ? null : instanceQuestionId,
       variant_id: variantId,
+      user_id: userId,
       variant_token: variantToken,
       submission_id: submissionId,
       url_prefix: urlPrefix,