+++ apps/prairielearn/src/lib/features/index.ts
@@ -6,6 +6,8 @@ const featureNames = [
   // Can only be applied to courses/institutions.
   'process-questions-in-worker',
   'question-sharing',
+  'bot-grading',
+  'disable-public-workspaces',
   // Can only be applied to institutions.
   'lti13',
   'terms-clickthrough',