+++ packages/formatter/src/date.ts
@@ -32,3 +32,43 @@ export function formatDate(
   }
   return dateFormatted;
 }
+
+/**
+ * Format a date to a human-readable string like '2020-03-27'.
+ *
+ * @param date The date to format.
+ * @param timeZone The time zone to use for formatting.
+ * @returns Human-readable string representing the date.
+ */
+export function formatDateYMD(date: Date, timeZone: string): string {
+  const options: Intl.DateTimeFormatOptions = {
+    timeZone,
+    hourCycle: 'h23',
+    year: 'numeric',
+    month: '2-digit',
+    day: '2-digit',
+  };
+  const parts = keyBy(new Intl.DateTimeFormat('en-US', options).formatToParts(date), (x) => x.type);
+  return `${parts.year.value}-${parts.month.value}-${parts.day.value}`;
+}
+
+/**
+ * Format a date to a human-readable string like '2020-03-27 14:27'.
+ *
+ * @param date The date to format.
+ * @param timeZone The time zone to use for formatting.
+ * @returns Human-readable string representing the date.
+ */
+export function formatDateYMDHM(date: Date, timeZone: string): string {
+  const options: Intl.DateTimeFormatOptions = {
+    timeZone,
+    hourCycle: 'h23',
+    year: 'numeric',
+    month: '2-digit',
+    day: '2-digit',
+    hour: '2-digit',
+    minute: '2-digit',
+  };
+  const parts = keyBy(new Intl.DateTimeFormat('en-US', options).formatToParts(date), (x) => x.type);
+  return `${parts.year.value}-${parts.month.value}-${parts.day.value} ${parts.hour.value}:${parts.minute.value}`;
+}