+++ packages/browser-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @prairielearn/browser-utils
 
+## 2.1.0
+
+### Minor Changes
+
+- d28a15d: Support custom return type for parseHTMLElement
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [0f7c90f]
+  - @prairielearn/html@4.0.2
+
 ## 2.0.1
 
 ### Patch Changes