+++ packages/html-ejs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prairielearn/html-ejs
 
+## 2.0.2
+
+### Patch Changes
+
+- 0f7c90f: Upgrade all JavaScript dependencies
+- Updated dependencies [0f7c90f]
+  - @prairielearn/html@4.0.2
+
 ## 2.0.1
 
 ### Patch Changes