+++ packages/postgres/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/postgres
 
+## 2.1.0
+
+### Minor Changes
+
+- 1de33d9: Add `formatQueryWithErrorPosition` function
+
+## 2.0.3
+
+### Patch Changes
+
+- 0f7c90f: Upgrade all JavaScript dependencies
+
 ## 2.0.2
 
 ### Patch Changes